chore: refactor form component and change export
