--- src/components/form/form.module.scss
@@ -1,30 +1,0 @@
-@import url(../../styles/global.scss);
-
-.formItem {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-}
-
-
-.formDescription { 
-  font-family: var(--font-family);
-  font-size: 0.875rem;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 20px;
-}
-
-
-.textDestructive{ 
-  color: var(--destructive);
-  font-family: var(--font-family);
-  font-size: 0.875rem;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 0.875rem;
-}
-
-.formMessage { 
-  font-size: 0.75rem;
-}
