+++ src/components/form/formDescription/formDescription.module.scss
@@ -1,0 +1,7 @@
+.formDescription {
+  font-family: var(--font-family);
+  font-size: 0.875rem;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 20px;
+}
