+++ src/components/form/formItem/formItem.module.scss
@@ -1,0 +1,5 @@
+.formItem {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
