+++ src/components/form/formLabel/formLabel.module.scss
@@ -1,0 +1,8 @@
+.textDestructive{
+  color: var(--destructive);
+  font-family: var(--font-family);
+  font-size: 0.875rem;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 0.875rem;
+}
