+++ src/components/form/formMessage/formMessage.module.scss
@@ -1,0 +1,3 @@
+.formMessage {
+  font-size: 0.75rem;
+}
